backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
django-filter==23.3
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.6.0

